refactor(contracts): Use uint256 for block numbers

+++ beacon-light-client/solidity/contracts/validators_accumulator/interfaces/IValidatorsAccumulator.sol
@@ -16,6 +16,6 @@ interface IValidatorsAccumulator {
   ) external payable;
 
   function findAccumulatorByBlock(
-    uint64 blockNumber
+    uint256 blockNumber
   ) external view returns (bytes32 accumulator);
 }
